pyproject.toml
[project]
name = "tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faker>=40.8.0",
    "pillow>=12.1.1",
    "stario>=2.2.1",
    "watchfiles>=1.1.1",
]
